ещё чёто делал

+++ src/RuntimeDTO.cs
@@ -1,3 +1,3 @@
 using System.Collections.Generic;
 
-public record RuntimeDTO(string stdout, string stderr, int exitCode, double runTime, double runRAM);
+public record RuntimeDTO(string stdout, string stderr, int exitCode, long runTime, long runRAM);
